followed comments to move some basic traits related functions to system.go
